+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # random-list Changelog
 
+## next
+
+- Changed text input focus behaviour to select all text on focus.
+
 ## 1.0.0 (2022-05-01)
 
 Initial release